project/frontend/requirements.txt
streamlit
torchvision
requests
pillow
numpy